Extract storage diffing/IO, gate validation and .gdd save behind preferences, restructure tests
