+++ desktop/wrapper/src/intercept_frontend_message.rs
@@ -37,13 +37,16 @@ pub(super) fn intercept_frontend_message(dispatcher: &mut DesktopWrapperMessageD
 			if let Some(path) = path {
 				dispatcher.respond(DesktopFrontendMessage::WriteFile { path, content });
 			} else {
+				// Derive the dialog filter from the suggested filename's extension so it tracks the
+				// editor's save-format preference (.gdd container or plain .graphite).
+				let extension = std::path::Path::new(&name).extension().and_then(|extension| extension.to_str()).unwrap_or("graphite").to_string();
 				dispatcher.respond(DesktopFrontendMessage::SaveFileDialog {
 					title: "Save Document".to_string(),
 					default_filename: name,
 					default_folder: folder,
 					filters: vec![FileFilter {
 						name: "Graphite Document".to_string(),
-						extensions: vec!["gdd".to_string()],
+						extensions: vec![extension],
 					}],
 					context: SaveFileDialogContext::Document { document_id, content },
 				});
