+++ editor/src/messages/portfolio/portfolio_message_handler.rs
@@ -14,6 +14,7 @@ use crate::messages::portfolio::document::graph_operation::utility_types::Transf
 use crate::messages::portfolio::document::node_graph::document_node_definitions;
 use crate::messages::portfolio::document::utility_types::network_interface::OutputConnector;
 use crate::messages::portfolio::document_migration::*;
+use crate::messages::portfolio::document_storage_io::{build_or_open_working_copy, compare_storage_against_runtime, open_gdd_document_future};
 use crate::messages::portfolio::utility_types::FileContent;
 use crate::messages::preferences::SelectionMode;
 use crate::messages::prelude::*;
@@ -208,8 +209,9 @@ impl MessageHandler<PortfolioMessage, PortfolioMessageContext<'_>> for Portfolio
 			}
 			PortfolioMessage::AutoSaveDocument { document_id } => {
 				let Some(byte_store) = resource_storage.storage() else { return };
+				let validate = preferences.validate_storage_round_trip;
 				let Some(document) = self.documents.get_mut(&document_id) else { return };
-				document.commit_storage_snapshot(byte_store);
+				document.commit_storage_snapshot(byte_store, validate);
 				responses.add(PersistentStateMessage::WriteDocument {
 					document_id,
 					document: document.serialize_document(),
@@ -316,7 +318,7 @@ impl MessageHandler<PortfolioMessage, PortfolioMessageContext<'_>> for Portfolio
 				// already holds the persisted cursor; re-committing here would stack a spurious interaction on it
 				// and make the first undo a no-op, so trust the restored cursor and skip the commit.
 				if !reopened && let Some(byte_store) = resource_storage.storage() {
-					document.commit_storage_snapshot(byte_store);
+					document.commit_storage_snapshot(byte_store, preferences.validate_storage_round_trip);
 					// Seal the freshly-converted document as the base revision: the conversion is a completed
 					// import, not an in-progress interaction, so retire it into history rather than leaving it in
 					// the hot log (where it would never reach an interaction boundary on a save without edits).
@@ -548,7 +550,7 @@ impl MessageHandler<PortfolioMessage, PortfolioMessageContext<'_>> for Portfolio
 					responses.add(NavigationMessage::CanvasPan { delta: (0., 0.).into() });
 				}
 
-				self.load_document(new_document, document_id, resource_storage, responses);
+				self.load_document(new_document, document_id, resource_storage, preferences.validate_storage_round_trip, responses);
 				responses.add(PortfolioMessage::SelectDocument { document_id });
 			}
 			PortfolioMessage::MoveAllPanelTabs {
@@ -789,6 +791,7 @@ impl MessageHandler<PortfolioMessage, PortfolioMessageContext<'_>> for Portfolio
 					document_path,
 					content,
 					store_handle,
+					preferences.validate_storage_round_trip,
 				));
 			}
 			PortfolioMessage::GddDocumentLoaded {
@@ -814,7 +817,7 @@ impl MessageHandler<PortfolioMessage, PortfolioMessageContext<'_>> for Portfolio
 				document.path = document_path;
 
 				// The working copy is already mounted (we opened the .gdd), so skip the async re-mount.
-				self.load_document(document, document_id, resource_storage, responses);
+				self.load_document(document, document_id, resource_storage, preferences.validate_storage_round_trip, responses);
 				responses.add(PortfolioMessage::SelectDocument { document_id });
 			}
 			PortfolioMessage::GddUndoRedoRebuilt { document_id, had_oracle, interface } => {
@@ -827,7 +830,7 @@ impl MessageHandler<PortfolioMessage, PortfolioMessageContext<'_>> for Portfolio
 					return;
 				};
 
-				document.apply_gdd_cursor_rebuild(interface, had_oracle, responses);
+				document.apply_gdd_cursor_rebuild(interface, had_oracle, preferences.validate_storage_round_trip, responses);
 			}
 			PortfolioMessage::ToggleResetNodesToDefinitionsOnOpen => {
 				self.reset_node_definitions_on_open = !self.reset_node_definitions_on_open;
@@ -974,7 +977,7 @@ impl MessageHandler<PortfolioMessage, PortfolioMessageContext<'_>> for Portfolio
 				document.name = self.resolve_document_name(candidate_name, None);
 
 				// Load the document into the portfolio so it opens in the editor
-				self.load_document(document, document_id, resource_storage, responses);
+				self.load_document(document, document_id, resource_storage, preferences.validate_storage_round_trip, responses);
 
 				responses.add(AppWindowMessage::Focus);
 
@@ -1843,7 +1846,14 @@ impl PortfolioMessageHandler {
 		}
 	}
 
-	fn load_document(&mut self, mut new_document: DocumentMessageHandler, document_id: DocumentId, resource_storage: &ResourceStorageMessageHandler, responses: &mut VecDeque<Message>) {
+	fn load_document(
+		&mut self,
+		mut new_document: DocumentMessageHandler,
+		document_id: DocumentId,
+		resource_storage: &ResourceStorageMessageHandler,
+		validate: bool,
+		responses: &mut VecDeque<Message>,
+	) {
 		let is_new_document = !self.document_ids.contains(&document_id);
 		if is_new_document {
 			self.document_ids.push_back(document_id);
@@ -1881,7 +1891,13 @@ impl PortfolioMessageHandler {
 			.get(&document_id)
 			.map(|document| document.network_interface.document_network().clone())
 			.unwrap_or_default();
-		responses.add(Self::mount_document_storage(self.working_copy_root.clone(), document_id, legacy_network, resource_storage.resources()));
+		responses.add(Self::mount_document_storage(
+			self.working_copy_root.clone(),
+			document_id,
+			legacy_network,
+			resource_storage.resources(),
+			validate,
+		));
 	}
 
 	/// Build the `FutureMessage` that constructs (or reopens) the document's `Gdd` working copy and
@@ -1898,6 +1914,7 @@ impl PortfolioMessageHandler {
 		document_id: DocumentId,
 		legacy_network: graph_craft::document::NodeNetwork,
 		byte_store: Box<dyn graph_craft::application_io::resource::LoadResource>,
+		validate: bool,
 	) -> Message {
 		let path = working_copy_root.map(|root| root.join(format!("{:x}", document_id.0)));
 		let editor_version = crate::application::GRAPHITE_GIT_COMMIT_HASH.to_string();
@@ -1912,7 +1929,7 @@ impl PortfolioMessageHandler {
 				}
 			};
 
-			if reopened {
+			if validate && reopened {
 				compare_storage_against_runtime(&gdd, &legacy_network, byte_store.as_ref(), document_id).await;
 			}
 
@@ -2086,301 +2103,6 @@ fn display_name_with_fallback(info: &DocumentInfo) -> String {
 		format!("Untitled Document ({:x})", info.id.0)
 	} else {
 		info.name.clone()
-	}
-}
-
-/// Build a document's `Gdd` working copy. With `Some(path)` it opens an existing on-disk working
-/// copy (reopen / autosave recovery) or creates a fresh one bound to `peer`, choosing the backend
-/// per platform: a folder on native, an OPFS directory on web (where `path` is the OPFS directory
-/// name). With `None` it creates a fresh in-memory working copy (tests, headless) that persists
-/// nowhere durable.
-///
-/// Returns the working copy plus whether it was reopened from an existing on-disk working copy
-/// (`true`) versus freshly created (`false`). Only a reopen has independently-read stored state and
-/// an embedded legacy blob worth comparing against the legacy load (compare-on-open).
-async fn build_or_open_working_copy(
-	path: Option<&std::path::Path>,
-	peer: graph_storage::PeerId,
-	document_uuid: u64,
-	version: String,
-) -> Result<(document_format::GddV1, bool), document_format::Error> {
-	use document_container::AnyContainer;
-	use document_format::{GddV1, GddV1Layout};
-
-	let Some(path) = path else {
-		let container = AnyContainer::Memory(document_container::backends::memory::MemoryBackend::new());
-		let gdd = GddV1::create_in(container, GddV1Layout, peer, document_uuid, version.clone(), version).await?;
-		return Ok((gdd, false));
-	};
-
-	#[cfg(not(target_family = "wasm"))]
-	let (container, exists) = {
-		use document_container::backends::folder::FolderBackend;
-		let exists = path.join("manifest.json").is_file();
-		let backend = if exists { FolderBackend::open(path)? } else { FolderBackend::create(path)? };
-		(AnyContainer::Folder(backend), exists)
-	};
-
-	#[cfg(target_family = "wasm")]
-	let (container, exists) = {
-		use document_container::AsyncContainer;
-		use document_container::backends::opfs::OpfsBackend;
-		let directory_name = path
-			.to_str()
-			.ok_or(document_format::Error::Container(document_container::ContainerError::InvalidPath(path.display().to_string())))?;
-		let backend = OpfsBackend::open(directory_name).await?;
-		let exists = backend.exists("manifest.json").await;
-		(AnyContainer::Opfs(backend), exists)
-	};
-
-	let gdd = if exists {
-		GddV1::open_in(container, GddV1Layout).await?
-	} else {
-		GddV1::create_in(container, GddV1Layout, peer, document_uuid, version.clone(), version).await?
-	};
-	Ok((gdd, exists))
-}
-
-/// Build the per-document container for a working copy at `path` (folder native, OPFS web), or an
-/// in-memory container when `path` is `None`. Does not open or create a `Gdd` — just the backend.
-async fn build_per_document_container(path: Option<&std::path::Path>) -> Result<document_container::AnyContainer, document_format::Error> {
-	use document_container::AnyContainer;
-
-	let Some(path) = path else {
-		return Ok(AnyContainer::Memory(document_container::backends::memory::MemoryBackend::new()));
-	};
-
-	#[cfg(not(target_family = "wasm"))]
-	{
-		use document_container::backends::folder::FolderBackend;
-		let backend = if path.join("manifest.json").is_file() {
-			FolderBackend::open(path)?
-		} else {
-			FolderBackend::create(path)?
-		};
-		Ok(AnyContainer::Folder(backend))
-	}
-
-	#[cfg(target_family = "wasm")]
-	{
-		use document_container::backends::opfs::OpfsBackend;
-		let directory_name = path
-			.to_str()
-			.ok_or(document_format::Error::Container(document_container::ContainerError::InvalidPath(path.display().to_string())))?;
-		let backend = OpfsBackend::open(directory_name).await?;
-		Ok(AnyContainer::Opfs(backend))
-	}
-}
-
-/// Build the `FutureMessage` that opens a `.gdd` archive: materialize it into the per-document working
-/// copy, build the runtime from the stored registry, validate it against the embedded legacy blob, and
-/// deliver the finished document via [`PortfolioMessage::GddDocumentLoaded`]. Resource bytes carried in
-/// the archive are extracted into `store_handle` (the app-global cache) so declarations and the runtime
-/// resolve. The registry build is authoritative; the embedded legacy blob is the always-checked oracle
-/// and the fallback if the build fails.
-fn open_gdd_document_future(
-	working_copy_root: Option<std::path::PathBuf>,
-	document_id: DocumentId,
-	document_name: Option<String>,
-	document_path: Option<std::path::PathBuf>,
-	content: Vec<u8>,
-	store_handle: crate::messages::resource_storage::ResourcesHandle,
-) -> Message {
-	let path = working_copy_root.map(|root| root.join(format!("{:x}", document_id.0)));
-
-	let future = async move {
-		let document = build_document_from_gdd(path.as_deref(), &content, &store_handle, document_id).await;
-		Message::Portfolio(PortfolioMessage::GddDocumentLoaded {
-			document_id,
-			document_name,
-			document_path,
-			document: document.map(Box::new),
-		})
-	};
-	future.into()
-}
-
-/// Build the `FutureMessage` that rebuilds a document's interface from its `Gdd` undo/redo cursor. The
-/// cursor already moved and persisted synchronously in the handler; `gdd` is a snapshot clone at that
-/// post-move position. This loads the cursor's declarations, converts the registry back to a runtime
-/// network with full metadata, builds the interface, and delivers it via [`PortfolioMessage::GddUndoRedoRebuilt`]
-/// so the handler can swap it in (preserving transients) and compare against the legacy oracle. `interface`
-/// is `None` on failure (logged here), in which case the handler leaves the live document untouched.
-pub(crate) fn rebuild_gdd_cursor_future(gdd: document_format::GddV1, store_handle: crate::messages::resource_storage::ResourcesHandle, document_id: DocumentId, had_oracle: bool) -> Message {
-	use crate::messages::portfolio::document::utility_types::network_interface::storage_metadata::build_interface_from_storage;
-
-	let future = async move {
-		let declarations = gdd.declarations(&store_handle).await;
-		let interface = match gdd.registry().to_runtime_with_full_metadata(&declarations) {
-			Ok((network, node_entries, network_entries)) => match build_interface_from_storage(network, node_entries, network_entries) {
-				Ok(interface) => Some(Box::new(interface)),
-				Err(error) => {
-					log::error!("Gdd undo/redo rebuild for {document_id:?}: failed to build interface: {error}");
-					None
-				}
-			},
-			Err(error) => {
-				log::error!("Gdd undo/redo rebuild for {document_id:?}: failed to convert registry to runtime: {error}");
-				None
-			}
-		};
-
-		Message::Portfolio(PortfolioMessage::GddUndoRedoRebuilt { document_id, had_oracle, interface })
-	};
-	future.into()
-}
-
-/// Core of the `.gdd` open: archive -> working copy -> `Gdd` -> runtime interface, with the embedded
-/// legacy blob loaded for validation (always compared) and fallback (used if the registry build fails).
-/// Returns the built document, or `None` if neither the registry build nor the legacy fallback worked.
-async fn build_document_from_gdd(
-	path: Option<&std::path::Path>,
-	content: &[u8],
-	store_handle: &impl graph_craft::application_io::resource::ResourceStorage,
-	document_id: DocumentId,
-) -> Option<DocumentMessageHandler> {
-	use crate::messages::portfolio::document::utility_types::network_interface::storage_metadata::{apply_network_view_settings, build_interface_from_storage, network_ids_from_entries};
-	use document_format::{GddV1, GddV1Layout};
-
-	let container = match build_per_document_container(path).await {
-		Ok(container) => container,
-		Err(error) => {
-			log::error!("Opening .gdd for {document_id:?}: failed to build working copy container: {error}");
-			return None;
-		}
-	};
-
-	let gdd = match GddV1::open_from_archive(content, container, GddV1Layout).await {
-		Ok(gdd) => gdd,
-		Err(error) => {
-			log::error!("Opening .gdd for {document_id:?}: failed to open archive: {error}");
-			return None;
-		}
-	};
-
-	// Extract resource bytes carried in the archive into the global cache so declarations + runtime resolve.
-	match gdd.resource_hashes().await {
-		Ok(hashes) => {
-			for hash in hashes {
-				match gdd.read_resource(&hash).await {
-					Ok(holder) => {
-						store_handle.store(holder.as_slice());
-					}
-					Err(error) => log::error!("Opening .gdd for {document_id:?}: failed to read resource {hash}: {error}"),
-				}
-			}
-		}
-		Err(error) => log::error!("Opening .gdd for {document_id:?}: failed to list resources: {error}"),
-	}
-
-	// Full legacy load of the embedded blob: serves as both the validation oracle and the fallback.
-	let legacy_document = gdd
-		.read_legacy_document()
-		.await
-		.and_then(|holder| String::from_utf8(holder.as_slice().to_vec()).ok())
-		.and_then(|serialized| DocumentMessageHandler::deserialize_document(&document_migration_string_preprocessing(serialized)).ok());
-
-	// Build the runtime from the stored registry (authoritative).
-	let declarations = gdd.declarations(store_handle).await;
-	let interface = match gdd.registry().to_runtime_with_full_metadata(&declarations) {
-		Ok((network, node_entries, network_entries)) => {
-			let network_ids = network_ids_from_entries(&network_entries);
-			match build_interface_from_storage(network, node_entries, network_entries) {
-				Ok(mut interface) => {
-					// Restore the per-network, per-peer view state (node-graph nav + previewing) from
-					// `session.json`; it travels in the archive but isn't in the registry/history.
-					apply_network_view_settings(&mut interface, &network_ids, gdd.network_view_settings());
-					Some(interface)
-				}
-				Err(error) => {
-					log::error!("Opening .gdd for {document_id:?}: failed to build interface: {error}");
-					None
-				}
-			}
-		}
-		Err(error) => {
-			log::error!("Opening .gdd for {document_id:?}: failed to convert registry to runtime: {error}");
-			None
-		}
-	};
-
-	// Compare the registry build against the legacy oracle (always, when both are available). Normalize the
-	// benign generic-vs-resolved `call_argument` difference (see `normalize_call_arguments_for_compare`) so
-	// it doesn't flag.
-	if let Some(interface) = interface {
-		if let Some(legacy) = &legacy_document {
-			let mut built = interface.document_network().clone();
-			let mut oracle = legacy.network_interface.document_network().clone();
-			normalize_call_arguments_for_compare(&mut built);
-			normalize_call_arguments_for_compare(&mut oracle);
-			if built != oracle {
-				log::error!(
-					"Open-.gdd divergence for {document_id:?}: registry-built document differs from the embedded legacy blob\n{}",
-					crate::messages::portfolio::document::diff_networks(&oracle, &built)
-				);
-			}
-		}
-		return Some(DocumentMessageHandler::from_storage(interface, gdd, String::new(), None));
-	}
-
-	// Registry build failed: fall back to the embedded legacy document.
-	log::warn!("Opening .gdd for {document_id:?}: registry build failed, falling back to embedded legacy document");
-	if legacy_document.is_none() {
-		log::error!("Opening .gdd for {document_id:?}: no embedded legacy document to fall back to");
-	}
-	legacy_document
-}
-
-/// Compare-on-open: convert the reopened `.gdd`'s stored registry back to a runtime network and diff
-/// it against `legacy_network` (the document the user actually loaded). Validates the `.gdd` read path
-/// against the legacy load. Legacy is authoritative during the soak, so this only logs divergence; it
-/// never swaps the candidate in. Cold-path (runs once per open, off the edit hot path).
-async fn compare_storage_against_runtime(
-	gdd: &document_format::GddV1,
-	legacy_network: &graph_craft::document::NodeNetwork,
-	byte_store: &dyn graph_craft::application_io::resource::LoadResource,
-	document_id: DocumentId,
-) {
-	let declarations = gdd.declarations(byte_store).await;
-	let mut candidate = match gdd.registry().to_runtime_with_metadata(&declarations) {
-		Ok((network, _entries)) => network,
-		Err(error) => {
-			log::error!("Compare-on-open for {document_id:?}: .gdd registry failed to convert to runtime: {error}");
-			return;
-		}
-	};
-
-	// One open/restore path resolves a node's generic `call_argument` (e.g. a layer's inner monitor node's
-	// `T`) to a concrete `Context`, while storage faithfully keeps the authored `T`. That generic-vs-resolved
-	// difference is benign (it's recomputed during compilation), so normalize it on both sides before
-	// comparing to keep the soak signal focused on real drift.
-	let mut legacy = legacy_network.clone();
-	normalize_call_arguments_for_compare(&mut legacy);
-	normalize_call_arguments_for_compare(&mut candidate);
-
-	if candidate != legacy {
-		log::error!(
-			"Compare-on-open divergence for {document_id:?}: the reopened .gdd does not match the legacy load\n{}",
-			crate::messages::portfolio::document::diff_networks(&legacy, &candidate)
-		);
-	}
-}
-
-/// Canonicalize every node's `call_argument` (recursively, including nested networks) so a generic `T`
-/// and its compilation-resolved concrete `Context` compare equal. Storage keeps the authored generic
-/// form, but one load/restore path resolves it to `Context`; that difference is benign for the
-/// compare-on-open soak check. Concrete `call_arguments` other than `Context` are left untouched.
-fn normalize_call_arguments_for_compare(network: &mut graph_craft::document::NodeNetwork) {
-	use graph_craft::Type;
-
-	let context = graph_craft::concrete!(graphene_std::Context);
-	for node in network.nodes.values_mut() {
-		if node.call_argument == context {
-			node.call_argument = Type::Generic(std::borrow::Cow::Borrowed("T"));
-		}
-		if let graph_craft::document::DocumentNodeImplementation::Network(inner) = &mut node.implementation {
-			normalize_call_arguments_for_compare(inner);
-		}
 	}
 }
