+++ editor/src/test_utils.rs
@@ -28,6 +28,10 @@ impl EditorTestUtils {
 
 		editor.handle_message(PortfolioMessage::Init);
 
+		// Enable the dual-write soak validation in tests so storage round-trip drift fails loud, matching
+		// the previous always-on-in-debug behavior now that it is gated by a preference.
+		editor.handle_message(PreferencesMessage::ValidateStorageRoundTrip { enabled: true });
+
 		Self { editor, runtime }
 	}
